design
This is the basic design. I attempted to get all the layouts complete
but I am still working on them I want everyone to have access to what I
have so far so I am committing the changes. I currently left off on the
last layout where we will display the clinic data with the map fragment.
Everything is not fully functional but this is a starting point.